+++ mediagrains/grains/CodedAudioGrain.py
@@ -11,10 +11,10 @@ from ..typing import (
     GrainDataParameterType)
 
 from ..cogenums import CogAudioFormat
-from .BaseGrain import BaseGrain
+from .Grain import Grain
 
 
-class CodedAudioGrain(BaseGrain):
+class CodedAudioGrain(Grain):
     """\
 A class representing a coded audio grain.
 
@@ -25,7 +25,7 @@ Any grain can be freely cast to a tuple:
 where meta is a dictionary containing the grain metadata, and data is a
 bytes-like object containing the coded audio data.
 
-The BaseGrain class provides a number of properties which can be used to access
+The Grain class provides a number of properties which can be used to access
 parts of the standard grain metadata, and this class inherits these:
 
 meta
