+++ mediagrains/grains/__init__.py
@@ -4,8 +4,7 @@ from .VideoGrain import VideoGrain
 from .CodedAudioGrain import CodedAudioGrain
 from .CodedVideoGrain import CodedVideoGrain
 from .EventGrain import EventGrain
-from .Grain import Grain
-from .BaseGrain import BaseGrain, attributes_for_grain_type, new_attributes_for_grain_type
+from .Grain import Grain, attributes_for_grain_type, new_attributes_for_grain_type, GrainFactory
 
 # Provide aliases for backwards compatibility with older versions that used separate grain constructors
 
@@ -15,9 +14,8 @@ CODEDAUDIOGRAIN: TypeAlias = CodedAudioGrain
 CODEDVIDEOGRAIN: TypeAlias = CodedVideoGrain
 EVENTGRAIN: TypeAlias = EventGrain
 GRAIN: TypeAlias = Grain
-BASEGRAIN: TypeAlias = BaseGrain
 
 
-__all__ = ["Grain", "VideoGrain", "CodedVideoGrain", "AudioGrain", "CodedAudioGrain", "EventGrain", "BaseGrain",
-           "GRAIN", "VIDEOGRAIN", "CODEDVIDEOGRAIN", "AUDIOGRAIN", "CODEDAUDIOGRAIN", "EVENTGRAIN", "BASEGRAIN",
-           "attributes_for_grain_type", "new_attributes_for_grain_type", "size_for_audio_format"]
+__all__ = ["Grain", "VideoGrain", "CodedVideoGrain", "AudioGrain", "CodedAudioGrain", "EventGrain",
+           "GRAIN", "VIDEOGRAIN", "CODEDVIDEOGRAIN", "AUDIOGRAIN", "CODEDAUDIOGRAIN", "EVENTGRAIN",
+           "attributes_for_grain_type", "new_attributes_for_grain_type", "size_for_audio_format", "GrainFactory"]
