+++ mediagrains/tools/wrap_in_gsf.py
@@ -26,7 +26,7 @@ from mediatimestamp.immutable import Timestamp
 
 from ..cogenums import CogFrameFormat, CogAudioFormat
 from ..grains import VideoGrain, CodedVideoGrain, AudioGrain, CodedAudioGrain
-from ..grains import BaseGrain
+from ..grains import Grain
 from ..gsf import GSFEncoder
 from ..utils import GrainWrapper, H264GrainWrapper, ADTSAACGrainWrapper
 from ._file_or_pipe import file_or_pipe
@@ -35,7 +35,7 @@ from ._file_or_pipe import file_or_pipe
 def wrap_to_gsf(
         input_file: str,
         output_file: str,
-        template_grain: BaseGrain,
+        template_grain: Grain,
         Wrapper: Union[Type[GrainWrapper], Type[H264GrainWrapper]] = GrainWrapper):
     """Wrap the supplied input in GSF and write it out to a given file-like object
