+++ static-commontooling/make/include/gitignore.mk
@@ -4,9 +4,10 @@
 
 # Custom gitignore file support
 ifeq "$(CUSTOM_GITIGNORE_FILE)" ""
-prepcode: $(topdir)/.gitignore
+$(topdir)/.gitignore: $(topdir)/.gitignore.gen
+	cp -f $< $@
 
-$(topdir)/.gitignore: $(commontooling_dir)/misc/$(CLOUDFIT_MAKE_MODE).gitignore
+$(topdir)/.gitignore.gen: $(commontooling_dir)/misc/$(CLOUDFIT_MAKE_MODE).gitignore
 	cp -f $< $@
 
 ifneq ($(EXTRA_GITIGNORE_LINES), "")
@@ -15,4 +16,25 @@ ifneq ($(EXTRA_GITIGNORE_LINES), "")
 		echo $$ignore_line >> $@ ; \
 	done
 endif
+
+check-gitignore: $(topdir)/.gitignore.gen
+	@diff $< $(topdir)/.gitignore || (\
+		echo "--------------------------------------------------------------------------------" && \
+		echo "" && \
+		echo ".gitignore is not identical to that generated by the commontooling templates." && \
+		echo "" && \
+		echo "Please run " && \
+		echo "" && \
+		echo "  make static-files" && \
+		echo "" && \
+		echo "and then check the new static files into git" && \
+		echo "" && \
+		echo "--------------------------------------------------------------------------------" && \
+		exit 1); \
+
+static-files: $(topdir)/.gitignore
+
+check-static-files: check-gitignore
+
+.PHONY: check-gitignore force_rebuild
 endif
