+++ static-commontooling/make/include/version.mk
@@ -3,17 +3,22 @@
 # Included by other .mk files, do not use directly.
 #
 
-PBRVERSION_VERSION?=1.2.0
+PBRVERSION_VERSION?=1.3.0
+PBRVERSION_CONTAINER?=bbcrd/pbrversion
 # Since the version is pinned above, there's no need to pull images every time, especially since this tool runs very
 # frequently
-PBRVERSION?=$(DOCKER) run --rm -v $(project_root_dir):/data:ro bbcrd/pbrversion:$(PBRVERSION_VERSION)
+PBRVERSION?=$(DOCKER) run --rm -v $(project_root_dir):/data:ro $(PBRVERSION_CONTAINER):$(PBRVERSION_VERSION)
 
 # If VERSION is't already set (because it was exported from a higher-level Makefile, extractit from a top level VERSION
 # file if present, otherwise we use PBR to extract the version from git in the right formats
 ifndef VERSION
-export VERSION := $(shell [ -f VERSION ] && cat VERSION || $(PBRVERSION))
-export DOCKERISED_VERSION := $(shell [ -f VERSION ] && cat VERSION || $(PBRVERSION) --docker)
-export NEXT_VERSION := $(shell [ -f VERSION ] && cat VERSION || $(PBRVERSION) --brief)
+# pbrversion --list gives an output of the form '<full version> <brief version> <dockerised version> <docker tag> [<docker tag>]'
+# which gets split out here into the four variables we need (while only calling pbrversion once).
+PBRVERSION_OUTPUT := $(shell $(PBRVERSION) --list)
+export VERSION := $(shell [ -f VERSION ] && cat VERSION || echo $(word 1, $(PBRVERSION_OUTPUT)))
+export DOCKERISED_VERSION := $(shell [ -f VERSION ] && cat VERSION || echo $(word 3, $(PBRVERSION_OUTPUT)))
+export NEXT_VERSION := $(shell [ -f VERSION ] && cat VERSION || echo $(word 2, $(PBRVERSION_OUTPUT)))
+export DOCKER_TAGS := $(wordlist 4, $(words $(PBRVERSION_OUTPUT)), $(PBRVERSION_OUTPUT))
 
 endif
